Systematics can be associated with a group of EDs.
The SystematicManager now stores its systematics in map of vectors.
Each vector is referred to as a group and is keyed by the name of that
group. Systematics are added to a group through the BinnedNLLH (lh)
like:

    lh.AddSystematic(Systematic, "groupName");

The order in which the systematic is added to a group is the order in
which it is applied. If a group name is not specified, the systematic is
added the "default" group which is applied to all EDs.

A ED is then associated with a vector of groups that will be applied to
that ED in that order:

    std::vector<std::string> appliedSys;
    appliedSys.push_back("GroupA");
    appliedSys.push_back("GroupB");
    lh.AddDist(ED,"appliedSys")

The "default" group is prepanded to the appliedSys if it is not
specified already.

See examples/systematicFit.cpp for example fit.