Prior and PriorManager class.
Added a Prior class that applies a penalty term to a BinnedNLLH
depending on how the primary variable compares to the underlining
function governing the prior.

Priors can be set with and without dependance on other floating
varibles.

The PriorManager class looks after all the applied priors and is used to
add the penalty term to the likelihood.

See examples/PriorExample.cpp for example uses.